app/requirements.txt
fastapi
uvicorn
pydantic==1.10.11
aiomysql
python-dotenv
debugpy